+++ src/views/statusNode.ts
@@ -1,8 +1,10 @@
-import { ExtensionContext, TreeItem, TreeItemCollapsibleState } from 'vscode';
+import { commands, Disposable, ExtensionContext, TreeItem, TreeItemCollapsibleState, Uri } from 'vscode';
 import { ExplorerNode, ResourceType } from './explorerNode';
-import { GitService, GitUri } from '../gitService';
+import { GitService, GitStatus, GitUri } from '../gitService';
 import { StatusFilesNode } from './statusFilesNode';
 import { StatusUpstreamNode } from './statusUpstreamNode';
+
+let _eventDisposable: Disposable | undefined;
 
 export class StatusNode extends ExplorerNode {
 
@@ -30,8 +32,8 @@ export class StatusNode extends ExplorerNode {
             children.push(new StatusUpstreamNode(status, 'ahead', this.context, this.git));
         }
 
-        if (status.state.ahead || (status.files.length !== 0 && this.git.config.insiders)) {
-            const range = status.state.ahead
+        if (status.state.ahead || (status.files.length !== 0 && this.includeWorkingTree)) {
+            const range = status.upstream
                 ? `${status.upstream}..${status.branch}`
                 : undefined;
             children.push(new StatusFilesNode(status, range, this.context, this.git));
@@ -40,12 +42,26 @@ export class StatusNode extends ExplorerNode {
         return children;
     }
 
-    async getTreeItem(): Promise<TreeItem> {
+    private _status: GitStatus | undefined;
+
+    async getTreeItem(): Promise < TreeItem > {
         const status = await this.git.getStatusForRepo(this.uri.repoPath!);
         if (status === undefined) return new TreeItem('No repo status');
 
+        if (_eventDisposable !== undefined) {
+            _eventDisposable.dispose();
+            _eventDisposable = undefined;
+        }
+
+        if (this.includeWorkingTree) {
+            this._status = status;
+
+            _eventDisposable = this.git.onDidChangeFileSystem(this.onFileSystemChanged, this);
+            this.git.startWatchingFileSystem();
+        }
+
         let hasChildren = false;
-        const hasWorkingChanges = status.files.length !== 0 && this.git.config.insiders;
+        const hasWorkingChanges = status.files.length !== 0 && this.includeWorkingTree;
         let label = '';
         let iconSuffix = '';
         if (status.upstream) {
@@ -68,7 +84,7 @@ export class StatusNode extends ExplorerNode {
             }
         }
         else {
-            label = `${status.branch} ${hasWorkingChanges ? 'has uncommitted changes' : 'is clean'}`;
+            label = `${status.branch} ${hasWorkingChanges ? 'has uncommitted changes' : this.includeWorkingTree ? 'has no changes' : 'has nothing to commit'}`;
         }
 
         const item = new TreeItem(label, (hasChildren || hasWorkingChanges) ? TreeItemCollapsibleState.Expanded : TreeItemCollapsibleState.None);
@@ -80,5 +96,22 @@ export class StatusNode extends ExplorerNode {
         };
 
         return item;
+    }
+
+    private get includeWorkingTree(): boolean {
+        return this.git.config.gitExplorer.includeWorkingTree;
+    }
+
+    private async onFileSystemChanged(uri: Uri) {
+        const status = await this.git.getStatusForRepo(this.uri.repoPath!);
+
+        // If we haven't changed from having some working changes to none or vice versa then just refresh the node
+        // This is because of https://github.com/Microsoft/vscode/issues/34789
+        if (this._status !== undefined && status !== undefined &&
+            ((this._status.files.length === status.files.length) || (this._status.files.length > 0 && status.files.length > 0))) {
+            commands.executeCommand('gitlens.gitExplorer.refreshNode', this);
+        }
+
+        commands.executeCommand('gitlens.gitExplorer.refresh');
     }
 }